requirements.txt
_libgcc_mutex==0.1
_openmp_mutex==5.1
abseil-cpp==20211102.0
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.2.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow-cpp==8.0.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
aws-c-common==0.4.57
aws-c-event-stream==0.1.6
aws-checksums==0.1.9
aws-sdk-cpp==1.8.185
babel==2.11.0
backcall==0.2.0
backports==1.0
backports.functools_lru_cache==1.6.4
beautifulsoup4==4.11.1
blas==1.0
bleach==5.0.1
blinker==1.4
boost-cpp==1.73.0
bottleneck==1.3.5
brotli==1.0.9
brotli-bin==1.0.9
brotlipy==0.7.0
bzip2==1.0.8
c-ares==1.19.0
ca-certificates==2023.05.30
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
cryptography==38.0.1
cuda==11.7.1
cuda-cccl==11.7.91
cuda-command-line-tools==11.7.1
cuda-compiler==11.7.1
cuda-cudart==11.7.99
cuda-cudart-dev==11.7.99
cuda-cuobjdump==11.7.91
cuda-cupti==11.7.101
cuda-cuxxfilt==11.7.91
cuda-demo-suite==12.0.76
cuda-documentation==12.0.76
cuda-driver-dev==11.7.99
cuda-gdb==12.0.90
cuda-libraries==11.7.1
cuda-libraries-dev==11.7.1
cuda-memcheck==11.8.86
cuda-nsight==12.0.78
cuda-nsight-compute==12.0.0
cuda-nvcc==11.7.99
cuda-nvdisasm==12.0.76
cuda-nvml-dev==11.7.91
cuda-nvprof==12.0.90
cuda-nvprune==11.7.91
cuda-nvrtc==11.7.99
cuda-nvrtc-dev==11.7.99
cuda-nvtx==11.7.91
cuda-nvvp==12.0.90
cuda-runtime==11.7.1
cuda-sanitizer-api==12.0.90
cuda-toolkit==11.7.1
cuda-tools==11.7.1
cuda-visual-tools==11.7.1
cycler==0.11.0
dbus==1.13.18
decorator==5.1.1
defusedxml==0.7.1
dgl==1.1.0.cu118
docker-pycreds==0.4.0
entrypoints==0.4
expat==2.4.9
ffmpeg==4.3
fftw==3.3.9
flit-core==3.6.0
fontconfig==2.14.1
fonttools==4.25.0
freetype==2.12.1
frozenlist==1.3.3
gdb==11.2
gds-tools==1.5.0.59
gflags==2.2.2
giflib==5.2.1
gitdb==4.0.10
gitpython==3.1.29
glib==2.69.1
glog==0.5.0
gmp==6.2.1
gmpy2==2.1.2
gnutls==3.6.15
google-api-core==2.11.0
google-api-python-client==2.83.0
google-auth==2.17.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
grpc-cpp==1.46.1
grpcio==1.42.0
gst-plugins-base==1.14.0
gstreamer==1.14.0
httplib2==0.22.0
icu==58.2
idna==3.4
importlib-metadata==4.11.4
importlib_resources==5.10.1
intel-openmp==2021.4.0
ipykernel==5.5.5
ipython==7.33.0
ipython_genutils==0.2.0
jedi==0.18.2
jinja2==3.1.2
joblib==1.1.1
jpeg==9e
json5==0.9.5
jsonschema==4.17.3
jupyter_client==7.0.6
jupyter_core==4.11.2
jupyter_server==1.23.4
jupyterlab==3.5.2
jupyterlab_pygments==0.2.2
jupyterlab_server==2.17.0
kiwisolver==1.4.4
krb5==1.19.2
lame==3.100
lcms2==2.12
ld_impl_linux-64==2.38
lerc==3.0
libboost==1.73.0
libbrotlicommon==1.0.9
libbrotlidec==1.0.9
libbrotlienc==1.0.9
libclang==10.0.1
libcublas==11.10.3.66
libcublas-dev==11.10.3.66
libcufft==10.7.2.124
libcufft-dev==10.7.2.124
libcufile==1.5.0.59
libcufile-dev==1.5.0.59
libcurand==10.3.1.50
libcurand-dev==10.3.1.50
libcurl==7.87.0
libcusolver==11.4.0.1
libcusolver-dev==11.4.0.1
libcusparse==11.7.4.91
libcusparse-dev==11.7.4.91
libdeflate==1.8
libedit==3.1.20221030
libev==4.33
libevent==2.1.12
libffi==3.4.2
libgcc-ng==11.2.0
libgfortran-ng==11.2.0
libgfortran5==11.2.0
libgomp==11.2.0
libiconv==1.16
libidn2==2.3.2
libllvm10==10.0.1
libnghttp2==1.52.0
libnpp==11.7.4.75
libnpp-dev==11.7.4.75
libnvjpeg==11.8.0.2
libnvjpeg-dev==11.8.0.2
libpng==1.6.37
libpq==12.9
libprotobuf==3.20.3
libsodium==1.0.18
libssh2==1.10.0
libstdcxx-ng==11.2.0
libtasn1==4.16.0
libthrift==0.15.0
libtiff==4.4.0
libunistring==0.9.10
libuuid==1.41.5
libwebp==1.2.4
libwebp-base==1.2.4
libxcb==1.15
libxkbcommon==1.0.1
libxml2==2.9.14
libxslt==1.1.35
lz4-c==1.9.4
markdown==3.4.1
markupsafe==2.1.1
matplotlib==3.5.2
matplotlib-base==3.5.2
matplotlib-inline==0.1.6
mistune==2.0.4
mkl==2021.4.0
mkl-service==2.4.0
mkl_fft==1.3.1
mkl_random==1.2.2
mpc==1.1.0
mpfr==4.0.2
mpmath==1.2.1
multidict==6.0.2
munkres==1.1.4
nbclassic==0.4.8
nbclient==0.6.8
nbconvert==7.2.7
nbconvert-core==7.2.7
nbconvert-pandoc==7.2.7
nbformat==5.7.1
ncurses==6.3
nest-asyncio==1.5.6
nettle==3.7.3
networkx==2.2
notebook==6.5.2
notebook-shim==0.2.2
nsight-compute==2022.4.0.15
nspr==4.33
nss==3.74
numexpr==2.8.4
numpy==1.21.5
numpy-base==1.21.5
oauthlib==3.2.2
openh264==2.1.1
openssl==1.1.1u
opt-einsum==3.3.0
orc==1.7.4
packaging==22.0
pandas==1.3.5
pandoc==2.19.2
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pcre==8.45
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.3.0
pip==22.3.1
pkgutil-resolve-name==1.3.10
ply==3.11
progress==1.5
prometheus_client==0.15.0
promise==2.3
prompt-toolkit==3.0.36
protobuf==4.21.12
psutil==5.9.0
ptyprocess==0.7.0
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyg==2.2.0
pygments==2.13.0
pyjwt==2.4.0
pyopenssl==22.0.0
pyparsing==3.0.9
pyqt==5.15.7
pyqt5-sip==12.11.0
pyro-api==0.1.2
pyro-ppl==1.8.4
pyrsistent==0.18.0
pysocks==1.7.1
python==3.7.15
python-dateutil==2.8.2
python-fastjsonschema==2.16.2
python_abi==3.7
pytorch==1.13.1
pytorch-cluster==1.6.0
pytorch-cuda==11.7
pytorch-mutex==1.0
pytorch-scatter==2.1.0
pytorch-sparse==0.6.16
pytz==2022.7
pyyaml==6.0
pyzmq==19.0.2
qt-main==5.15.2
qt-webengine==5.15.9
qtwebkit==5.212
re2==2022.04.01
readline==8.2
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scikit-plot==0.3.7
scipy==1.7.3
seaborn==0.12.2
send2trash==1.8.0
sentry-sdk==1.12.1
setproctitle==1.3.2
setuptools==65.5.0
shortuuid==1.0.11
sip==6.6.2
six==1.16.0
smmap==5.0.0
snappy==1.1.9
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlite==3.40.0
sympy==1.10.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardx==2.2
terminado==0.17.1
threadpoolctl==2.2.0
tinycss2==1.2.1
tk==8.6.12
toml==0.10.2
tomli==2.0.1
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.1
tqdm==4.64.1
traitlets==5.8.0
typing-extensions==4.4.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
utf8proc==2.6.1
wandb==0.13.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
werkzeug==2.2.2
wheel==0.37.1
xz==5.2.8
yarl==1.8.1
zeromq==4.3.4
zipp==3.11.0
zlib==1.2.13
zstd==1.5.2
